🏗️ "Screaming Architecture"を採用し応用した構成に移行
