+++ frontend/src/features/common/ErrorAlert.tsx
