+++ frontend/src/features/common/Header.tsx
@@ -6,6 +6,7 @@ const Header = ({
   playToggle: () => void;
 }) => {
   const audioSwapText = playing ? "音楽を停止" : "音楽を再生";
+  
   return (
     <header className="navbar flex justify-between bg-gradient-to-r from-primary to-secondary">
       <div className="ml-4">
