+++ frontend/src/features/entry/index.ts
@@ -1,0 +1,4 @@
+import EntryForm from "./EntryForm";
+
+// ゲームに参加するための機能を提供するコンポーネントをまとめてエクスポート
+export { EntryForm };
