+++ frontend/src/features/play/NextGameLink.tsx
@@ -1,0 +1,59 @@
+import { socket } from "@/utils/socket";
+import SocketEvent from "@/class/socketEvents";
+
+const NextGameLink = ({
+  isHost,
+  roomId,
+}: {
+  isHost: boolean;
+  roomId: string;
+}) => {
+  // もう一戦ボタンを押したときの処理
+  const handleNextClick = () => {
+    const parameter = { roomId };
+    socket.emit(SocketEvent.REQ_NEXTGAME, JSON.stringify(parameter));
+  };
+
+  // ゲームを終了ボタンを押したときの処理
+  const handleQuitClick = () => {
+    const parameter = { roomId };
+    socket.emit(SocketEvent.REQ_CLOSEROOM, JSON.stringify(parameter));
+  };
+
+  if (!isHost) {
+    return null;
+  } else {
+    return (
+      <div className="flex gap-2 mt-4">
+        <button
+          className="btn btn-outline btn-primary"
+          onClick={handleNextClick}
+        >
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            fill="none"
+            viewBox="0 0 24 24"
+            strokeWidth={1.5}
+            stroke="currentColor"
+            className="size-6"
+          >
+            <path
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              d="M16.023 9.348h4.992v-.001M2.985 19.644v-4.992m0 0h4.992m-4.993 0 3.181 3.183a8.25 8.25 0 0 0 13.803-3.7M4.031 9.865a8.25 8.25 0 0 1 13.803-3.7l3.181 3.182m0-4.991v4.99"
+            />
+          </svg>
+          もう一戦！
+        </button>
+        <button
+          className="btn btn-outline btn-secondary"
+          onClick={handleQuitClick}
+        >
+          ゲームを終了する
+        </button>
+      </div>
+    );
+  }
+};
+
+export default NextGameLink;
