+++ frontend/src/features/play/index.ts
@@ -1,0 +1,22 @@
+import ThemeCard from "./ThemeCard";
+import NumberCard from "./NumberCard";
+import AnswerForm from "./AnswerForm";
+import SortAnswerBoard from "./SortAnswerBoard";
+import FinalAnswerBtn from "./FinalAnswerBtn";
+
+import FinalResultCard from "./FinalResultCard";
+import NextGameLink from "./NextGameLink";
+import ResultTable from "./ResultTable";
+
+
+// ゲームをプレイするための機能を提供するコンポーネントをまとめてエクスポート
+export {
+  ThemeCard,
+  NumberCard,
+  AnswerForm,
+  SortAnswerBoard,
+  FinalAnswerBtn,
+  FinalResultCard,
+  NextGameLink,
+  ResultTable,
+};
