+++ frontend/src/features/start/index.ts
@@ -1,0 +1,7 @@
+import IdCard from "./IdCard";
+import PlayerTable from "./PlayerTable";
+import DescAccordion from "./DescAccordion";
+import InputThemeForm from "./InputThemeForm";
+
+// ゲームを開始するための機能を提供するコンポーネントをまとめてエクスポート
+export { IdCard, PlayerTable, DescAccordion, InputThemeForm };
