+++ frontend/src/index.html
@@ -1,6 +1,20 @@
 <!DOCTYPE html>
 <html data-theme="pastel" lang="ja">
   <head>
+    <!-- meta -->
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta
+      name="description"
+      content="大人気パーティーゲームがブラウザ版アプリとして遊べるようになりました。ルールは1つ、数字を口に出したらアウト！テーマに沿ってその数字の大きさを表現し合い、配られた1~100のカードの並び順を協力して当ててクリアを目指します。"
+    />
+    <meta name="format-detection" content="telephone=no" />
+
+    <!-- favicon -->
+    <link rel="icon" href="favicon.ico" />
+    <link rel="apple-touch-icon" href="webclip.png" />
+
+    <!-- font -->
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
@@ -14,6 +28,28 @@
       href="https://fonts.googleapis.com/css2?family=Comic+Neue:wght@400;700&family=Kiwi+Maru:wght@400;500&display=swap"
       media="print"
       onload='this.media="all"'
+    />
+
+    <!-- ogp -->
+    <meta property="og:site_name" content="ito(イト)" />
+    <meta property="og:url" content="https://www.ito-game.com/" />
+    <meta property="og:type" content="website" />
+    <meta property="og:title" content="ito(イト) | 意思疎通ゲーム" />
+    <meta
+      property="og:description"
+      content="大人気パーティーゲームがブラウザ版アプリとして遊べるようになりました。ルールは1つ、数字を口に出したらアウト！テーマに沿ってその数字の大きさを表現し合い、配られた1~100のカードの並び順を協力して当ててクリアを目指します。"
+    />
+    <meta property="og:image" content="https://www.ito-game.com/ogp.webp" />
+    <meta property="og:locale" content="ja_JP" />
+    <meta name="twitter:card" content="summary_large_image" />
+    <meta name="twitter:site" content="@yamap_web" />
+    <meta
+      name="twitter:description"
+      content="大人気パーティーゲームがブラウザ版アプリとして遊べるようになりました。ルールは1つ、数字を口に出したらアウト！テーマに沿ってその数字の大きさを表現し合い、配られた1~100のカードの並び順を協力して当ててクリアを目指します。"
+    />
+    <meta
+      name="twitter:image:src"
+      content="https://www.ito-game.com/ogp.webp"
     />
   </head>
   <body>
